[BP] MB-48759: End legacy DCP stream if there are no items in range [2/2]
Cherry-pick of ce8a925e

Currently if a legacy DCP stream sets a stream range from X to Y and
there are no items within this range for the default collection we
won't end the stream when streaming from in memory. We need to do
this as we can't just send a SnapshotMarker op, then a SeqnoAdvanced
op to take us the end of the stream. As legacy DCP clients won't be
aware of SeqnoAdvanced ops.

To fix this, add a check to ActiveStream::processItems() where if there
are no mutations for the stream and the last seqno we observed was
greater or equal to the stream range's end value then we should end the
stream. This should only be done if the stream is a legacy stream that
doesn't support sync-writes.

Change-Id: Ica8ae65ab7c658da3bcae8069a486522f169d042
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182376
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>