MB-48759: End legacy DCP stream if there are no items in range [2/2]
Currently if a legacy DCP stream sets a stream range from X to Y and
there are no items within this range for the default collection we
won't end the stream when streaming from in memory. We need to do
this as we can't just send a SnapshotMarker op, then a SeqnoAdvanced
op to take us the end of the stream. As legacy DCP clients won't be
aware of SeqnoAdvanced ops.

To fix this, add a check to ActiveStream::processItems() to end the
stream if highest process item in the checkpoint 'curChkSeqno' is equal
or greater than the stream's end seqno. Also remove a check from
ActiveStream::inMemoryPhase() that used to do this based on the
'lastSentSeqno' as it doesn't work in situation for collection only
streams. Also this would normally be called after processItems(), which
will end the stream before hand meaning inMemoryPhase() will not be
called again.

This patch also adds more unit tests for legacy streams and collections
streams. Checking they see end stream ops when streaming from memory
with a end stream seqno.

Change-Id: Ica8ae65ab7c658da3bcae8069a486522f169d042
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163597
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>