MB-48467: Add config for magma MinCheckpointCreationInterval
Magma will create a checkpoint if the overhead of the checkpoint
queue exceeds 10% of the data size including fragmentation. It is
configured to 2% per checkpoint via magma_checkpoint_threshold
in config.json which results in 10% since magma holds a max of
five checkpoints.

During data load or overwrite workloads where the data size is
small, this can result in agressive checkpoint creation. This
patch adds MinCheckpointCreationInterval config to rate limit
checkpoint creation in such workloads.

Also set the config to zero for tests which require creation of
a checkpoint every batch.

Change-Id: If09b7fd6d7615a583d8f99c050682c66b1842727
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163750
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>