MB-43238: Fix intermittent failure of expiry pager settings test
The test checks that the expected run time of the expiry pager task is
within an upper and lower bound based on the time before/after setting
the config param, plus the sleep time.

This ensures the test is resilient to time passing between computing an
expected time and the config actually being applied.

However, the test relies on system_clock, whereas the expiry pager
config derives the current time using ep_current_time(), which is
subject to memcached_uptime ticks.

The time computed by the expiry pager can then appear to be (at most)
one second behind system_clock, dependent on exactly when the time was
determined relative to when the tick occurs.

Fix by relaxing the test checks by one second.

A better fix might be to settle on one time source for the test and
pager for consistency, but relaxing the test is a less invasive path to
resolving the intermittent failure.

Change-Id: I53efc937f07a48ce1d3b2157ec3aea31efe93b38
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163884
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>