MB-48923: Avoid underflow of on_disk_prepare_bytes during compaction
During upgrade from <6.6.1 to 6.6.2+, a new stat is added and tracked
in _local/vbstate to count the bytes in on-disk Prepares - see
MB-42306.

There is a bug in how the updates to this stat are handled during
compaction - the value is not clamped to zero and hence can underflow:

    auto prepareBytes = json.find("on_disk_prepare_bytes");
    if (prepareBytes != json.end()) {
        auto onDiskPrepareBytes =
                std::stoull(prepareBytes->get<std::string>()) -
                ctx.stats.prepareBytesPurged;
        *prepareBytes = std::to_string(onDiskPrepareBytes);
        updateVbState = true;
    }

Note that if ctx.stats.prepareBytesPurged is greater than prepareBytes
read from disk, the value written will underflow to a negative number
- or very large positive if represented as a unsigned value.

This issue was fixed in 7.0.0 upwards as part of
http://review.couchbase.org/c/kv_engine/+/140801 when the
on_disk_prepare_bytes code was forward-merged from mad-hatter to
master; however the fix wasn't flagged as needing backporting at the
time.

This patch backports it to mad-hatter branch.

Change-Id: Ida02f5a4752bb6dc28a9ab2316acf69972a10608
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163890
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>