MB-49022: Always reset the SV committed state on expiry
Currently if we expire an item with xattrs then we enter
HashTable::unlocked_softDelete with onlyMarkDeleted as true and
skip the setting of the committed state to Mutation. Then, as we
attempt to enqueue the item into the CheckpointManager we fail an
assert that every commit must have a durability ctx in the
VBQueueItemCtx. We should reset the committed state in all cases
to Mutation.

Change-Id: I3cc6f989803877e6fdc2a0a2494ed85d3e96d5de
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163956
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>