MB-48915: Compute checkpoint-max-size on the exact num-vbuckets on node
Currently the computation accounts the capacity of the VBucketMap,
while we want an accurate computation on the actual number of
vbucket-objects managed by a node at any given time.

The way to achieve that is recomputing the new checkpoint-max-size every
time a vbucket is created or deleted in the VBucketMap.

Change-Id: I02a7797c09660eb3b7781a73eece57c8bfcc1c9a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/164131
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>