MB-47134: Ensure we don't underflow aggregated flush stats
Ensure we don't underflow when aggregating flush stats in a flush
failure situation. Due to us not account for de-duplication of items
correctly.

For instance, if we have two persisted keys A & B.
1. A:1, B:2
We then have a key C come into the checkpoint and try to persist it
2. A:1, B:2 [B], C:3 [P]
This fails but a new write comes in for key B
3. A:1 [B], C:3 [P], B:4
We then try and perform a flush again of the checkpoint, at which point
we throw as we underflow the dirtyQueueSize. As we incorrectly account
B:2 as being de-duplicated during the failed flush, by it having been
placed in persistenceFailureStatOvercounts.

To fix this we need to change the logic to add an item to
persistenceFailureStatOvercounts.
From:
backupPCursorSeqno <= existingSeqno
To:
existingSeqno > initialBackupPCursorSeqno
Where the backup seqno is taken before we start iterating though cursors
so it's the original seqno of the backup cursor. If the backup cursor is
pointing to a meta item then we dec the seqno to point to the previous
visible item.

Change-Id: I3fadabff4f41d44573af19a879b908139f94b509
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/164158
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>