MB-49040: 6/n Only check limits when limits are present
Avoid a scope map lookup + value compare if there are no limits
to check. A VB:Manifest flag now exists that is set everytime a
VB::Manifest is updated against a Manifest from the cluster.
If any scope is found to have a limit, we enable the flag or
disable if no scopes with limits are found.

Change-Id: I83a7062822d6eb7805164e411600a6287b3bf737
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/164342
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>