MB-48368: Generate correct seqno advance for in-memory stream
Another case missed in ActiveStream::processItems when a checkpoint
contains only a prepare/abort for the stream's collection.

This is fixed with an extra check in path when the output mutations
is empty - we now detect that the cursor visited a non-visible
item for the stream and generate an empty snapshot to catch the
client up to the collection high-seqno (so they can no be happy
they have a complete 'copy' of the collection).

Testing is enhanced in this patch to be much more data-driven
and simpler - targetting just ActiveStream::processItems with
different sequences of inputs and checking the outputs.
Without the fixes in ActiveStream a number of the new tests
detect problems with the stream's responses.

Change-Id: Id0fb4e83e8be00cbf1ffc4db4fccdf6dfd959abb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/163468
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>