MB-49207: Fix BgFetcher race restoring old items for full fetches
There exists a race in the BgFetcher that means that old versions
of items can be restored to the HashTable if an item is updated
and evicted after the BgFetcher has read the item from disk but
before it restores it to the HashTable.

This change does not include fixes for the meta only fetching path,
or compaction expiry bg fetching paths.

Change-Id: I33041a3e28e1b78a433ff42f4a345f104b106d37
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165999
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>