MB-48587: Introduce Checkpoint::queueMemOverhead
The new member accounts for the memory overhead of internal pointers of
the Checkpoint::toWrite container.

From the Pov of memory tracking in CM, we already have:
- Checkpoint::queuedItemsMemUsage: which account the sizes of all items
  queued into the checkpoint
- Checkpoint::keyIndexMemUsage: accounts the full allocation in the
  key-index (internal struct + keys allocation)
The new member is the missing component for the total CM mem usage.

Note that the new member type's MemoryCounter. Which means that with
this patch we start accounting the toWrite memory overhead into the
estimated CM mem usage at all levels (bucket, vbucekt/CM, checkpoint).
That makes the accounting for enforcing the CM Quota more accurate.

Change-Id: I34e05d88f57cfea9f2b68827bcde5bd031a9bf4d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/162237
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>