Make KVBucketTest::store_items() simpiler to use
Currently KVBucketTest::store_items() returns the status of if all
writes have been successful, which requires the caller to wrap the call
to store_items() in a ASSERT_TRUE()/EXPECT_TRUE(). However, if you don't
look at the definition of the function that this might not be clear. So
add [[nodiscard]] declaration to ensure that that caller checks the
return value.

Change-Id: I4a89fa24c2fcaf3476af9e3d42c816fd32941018
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/164992
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>