Change checkpoint_memory_ratio from 50% to 30%
The original purpose of capping the Checkpoint mem usage is to avoid
that it takes up the entire bucket quota and leave more memory available
for other components.

One of the components that most gains from that is the HashTable.
So, first the aim here is to run with a lower CM-quota default to verify
that all QE/Perf tests are not negatively affected by that. Also, we
want to verify the effect of this change on specific ResRatio/GetLatency
perf tests.

Change-Id: I7d1ac440d4883215b67238245d02a78c403c0522
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/164997
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>