MB-47318: Refactor: Move dcp_snapshot_marker_codec.h/cc
Move the code to a more accessible place so that unit tests of
MB-47318 can be backported to mad-hatter.

This commit is a recreation and simplification of cca87559 which
does a more extensive job of moving this code upstream. This
particular change is just for allowing the backports and can be
ignored in forward merging.

Change-Id: Icfcb85ca1147fac33da62fb00fc43641f197f155
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165013
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Well-Formed: Restriction Checker