MB-49348: Fix race in time based CAS
Two racing threads can currently receive the same CAS value when
attempting to generate a new CAS. Currently this can only be done
for different documnets as we always call this while holding a
HashBucketLock. Correct this race by checking the result of a
compare_exchange_strong when using the real clock based CAS and
repeating the CAS generation if not successful.

Change-Id: Ia4d0109065b4435bf21658d446a2710e124b5562
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/165101
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>