MB-49207: Use cas of StoredValue driving BgFetch as a completion token
A BgFetch should only be allowed to ovewrite the value in the
HashTable if that value is the one that originally drove the BgFetch.
This is because the item in the HashTable could have been replaced by
a new revision. Use the cas of the StoredValue driving the BgFetch
as a token that will be checked against the cas of the item at the
completion of the BgFetch to determine if it's valid to replace the
value in the HashTable with the value that we just fetched from disk.

Change-Id: If2ddc6123091e4fb2a00a6cd4c4e46ea22773bf7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165162
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>