MB-47318: [BP] Save the 'pristine' message size when consumer buffers
If the Consumer buffers a DcpResponse, release and store the
UpdateFlowControl value along size the DcpResponse. This value
is then used for calculating DCP buffer ACKs, not the possibly
incorrect DcpResponse::getMessageSize

This commit includes a fix to the originally merged change.

d83d53e6

MB-47318: Address double acking when closing a stream

When the DcpConsumerTask used 'moveFromFront' and took ownership
of DcpResponse, the Buffer::bytes member still accounted for
the size of the DcpResponse.

If a close stream occurred we would add Buffer::bytes to the
FlowControl ack size and then also if the DcpResponse was also
processed the DcpResponse size is also added - a double count.

This is fixed by ensuring Buffer::bytes only accounts for the
items owned by the Buffer::messages queue.

An existing test was already exercising this scenario, but lacked
the correct EXPECTS to catch - now fixed.

Change-Id: I3ea238d4b182153d39ab44e8acba4fe007437470
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166228
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>