MB-49388: Avoid repeated memmove in AsyncReadCallback
A regression in the time taken to retrieve `collections-details` stats
with `mcstats` was noted. For 1000 collections in a cluster run on a
2019 macbook, master took ~5 minutes, vs 7.0.2 ~30s.

Perf investigation showed that this was not a server side regression,
rather a regression in `mcstats` itself - CC `mcstat` against a master
node did not demonstrate the regression, while master `mcstat` against
CC node did.

Profiling found the a large amount of time was spent in memmove,
shuffling unread bytes to the start of a buffer in
AsyncReadCallback::drain.

To avoid this, switch to using folly:IOBuf, which handles head/tail
pointers into a buffer, and only memmove data to the start of the
buffer if absolutely necessary (or when reallocating the buffer).

This change brings the average (n=5) time taken to 25s - similar to the
pre-regression value.

Change-Id: Ic06336ce638ec1afed54b000e44a587173108e06
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/165159
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>