Add mcbasher
Add back a new mcbasher (built on folly this time) which generate
load to a memcached node

Usage:
    mcbasher --host localhost \
             --user myuser \
             --password mypassword \
             --bucket mybucket \
             --connections 100 \
             --threads 4
             --pipeline \
             --ooo

Creates 4 threads with 100 connections using OoO and pipelines
data to hammer the server.

Instead of --pipeline and --ooo one may use --disconnect which
insert disconnects into the mix of operations to send to the
server.

Change-Id: I6b3845f70da35f816de7d331a4b3683d2e579716
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165330
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>