MB-49207: Take rvalue HBL in bgFetch
In a future patch we'll pass along the temp StoredValue added for
this fetch and the HBL allows us to do so without it being changed
under us.

This patch moves all of the manual unlocks of the HBL before
queueing a BGFetch into one place (bgFetch()). It's not clear why
exactly this was done, but it's suspected that it's because we are
now trying to acquire a vBucket wide lock to queue the BgFetch.

Change-Id: Ic0c584e05b4fd8a4af1db2ec098cc996f5705510
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/165353
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>