MB-49465: Don't dereference nullptr in makeImplicitCompactionContext
Don't dereference a nullptr returned from makeCompactionContextCallback
in MagmaKVStore::makeImplicitCompactionContext(), as it will cause us to
crash. Instead return a nullptr to the caller. Also update
MagmaKVStore::compactionCallBack() to handle a nullptr and return false
it its caller.

This patch also adds a couple of tests to ensure that this change works
as expected.

Change-Id: Ifb380d64c85671a1284f8d20b1e675cd1480548d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165494
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>