MB-49472: Only remove a dropped collection once fully purged
When calculating the dropped collection disk metadata
Flush::encodeRelativeComplementOfDroppedCollections needs more
information from the "eraser". We pass in the end-seqno, which
is maximum end of all dropped collections from the pre-compaction
snapshot.

E.g. if a snapshot has the following list of dropped collections

* id:8, start:1, end:10
* id:9, start:3, end:15

The total extent of seqnos for dropped collections is 1 to 15, and
for this fix we're interested in 15.

If a collection is resurrected whilst compaction runs, it will be
created with a start > 15. If that resurrected collection is dropped
it will have an end > 15.

encodeRelativeComplementOfDroppedCollections now uses that value
(e.g. 15) to see if a collection should still be included in the
output.

For example if collection with id:8 was resurrected at 16 and dropped
at 18 during compaction encodeRelativeComplementOfDroppedCollections
is working with the following parameters.

droppedCollections
* id:8, start:1, end:18
* id:9, start:3, end:15

idsToRemove
* id:8
* id:9

endSeqno (new parameter in this commit)
* 15

Using this information we know that compaction has not reached the end of
collection with id:8, but has seen all of id:9.

The output is now:

* id:8, start:1, end:18

Before this fix, the output was empty.

Change-Id: I51f945364d62934a5d14f8d72940a4362d371611
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165470
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>