MB-49037: Add ep_ht_item_memory stat
For the HT metadata we expose both the vb_{state} and the EP
aggregators, while for data (ie, items) we expose only the vb_{state}
one.

ep_ht_item_memory is the EP total allocations of items in all HTs.

The patch also renames some existing stats with a more descriptive
name:
  vb_{state}_itm_memory -> vb_{state}_ht_item_memory
  vb_{state}_itm_memory_uncompressed -> vb_{state}_ht_item_memory_uncompressed

Change-Id: Ia9564ab58940f690a49e612fe538489faddb370f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165745
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>