MB-49542: Ensure OSO backfill does not register cursor for dead stream
ActiveStream removes any associated cursor when transitioning to dead.
After this, the stream should never again have a cursor registered, as
it will not be cleaned up when the stream is destroyed.

Backfills may attempt to re-register a cursor. Seqno ordered backfills
already checked that the stream is in the correct state; OSO backfills
lacked this check.

As a result, creating a stream, starting an OSO backfill, then ending
the stream could lead to the backfill task registering a cursor for a
dead stream.

This cursor would then linger forever; no stream would clean it up, and
cursor dropping would never remove it as that code path calls through
DcpProducer to find the matching stream - and there is no matching
stream.

Change-Id: Iba3d613eeec2f9a56327e5ca2a1f7793d1643ad7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165830
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>