MB-49526: Modify MagmaKVStore MagmaCompactionCB & compactionCallBack
Modify MagmaCompactionCB so that we create CompactionContext in the
ctor of MagmaCompactionCB. Rather than on the first call to
MagmaKVStore::compactionCallBack(). Also throw if we're unable to create
a CompactionContext, preventing Magma from running a compaction that
can't do anything.

A few unit tests are skipped when running with Nexus, as they trigger
Magma's implicit compaction. This causes issues as we currently don't
support implicit compaction when running under the Nexus back end,
meaning we don't have MagmaKVStore::makeCompactionContextCallback set,
which intern will cause us to throw in the MagmaCompactionCB()
constructor.

Change-Id: Ie4f0a176f51bbbdb5d55c3c41711aaac8f47bf15
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/165921
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>