MB-49526: Update MagmaCompactionCB to set it's magma::Status
Update MagmaCompactionCB so that it sets it's magma::Status code if it
encounters a failure, so that we can cancel the compaction. As it might
not be possible to continue with it if for instance the in memory
vbucket has been destroyed.

To enable this change update we've also modified compactionCallBack()
and compactionCore() of MagmaKVStore to return a pair of a magma::Status
and a bool if to drop the doc.

Change-Id: I7cf55279b56485ca721df2c61bfe320091f40566
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166362
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>