MB-49453: Don't generate a new snapshot when ending on a new checkpoint
An issue occurs if the following sequence is passed to
ActiveStream::processItems and the stream does not support sync
replication.

* 1:prepare
* 2:mutation
* checkpoint-end
* checkpoint-start

Prior to this commit the presence of 1:prepare lead to an extra, invalid
snapshot+seqno advance being sent when processItems reaches the
checkpoint-start (and sends the correct snapshot). We can prevent this
by clearing the std::optional "highNonVisibleSeqno" when we generate
each snapshot.

Change-Id: I576c5cdb4b2b5e9678e8156dbdf757adc97a6563
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166077
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>