MB-49036: Allow ReviveDocument to be used with durability
`subdoc_executor` ensures that a subdoc operation with ReviveDocument
set is not operating on a document which is Alive (not deleted).

However, as part of `subdoc_update`, the document state was overwritten
with Alive prior to performing an Add.

With a non-durable operation, this did not lead to issues; with a
durable op it is expected that the operation will ewouldblock, and
later be resumed.

Once the sync write has completed, `subdoc_executor` will be executed
again, and would fail the operation with

 SubdocCanOnlyReviveDeletedDocuments

as the state had been set to Alive.

Change-Id: I4fea5fd561c97be84deb9ce23bc37f6541d627a6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166233
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>