ep_testsuite: Avoid unnecessary exception throwing
Instead of unconditionally calling cb::io::rmrf() on the directory to
be used for this test; check it exists first. This avoids throwing an
exception (which is immediately caught and swallowed) if the directory
doesn't exist.

This simplifies running under a debugger where one wants to catch any
thrown exceptions (which are _normally_ exceptional, error cases).

Change-Id: Id4337a07f165e2d280ca3ccaa1e51e68a7c2a09e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167594
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>