KVBucketTest: Avoid unnecessary exception throwing
Instead of unconditionally calling cb::io::rmrf() on the directory to
be used for this test; check it exists first. This avoids throwing an
exception (which is immediately caught and swallowed) if the directory
doesn't exist.

This simplifies running under a debugger where one wants to catch any
thrown exceptions (which are _normally_ exceptional, error cases).

Change-Id: I06d7bfe694b739cb9a6b49c0615e7b8d74716510
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166355
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>